(!) update checkers structure
