+++ src/core/checkers/__init__.py
@@ -1,0 +1,1 @@
+from .max_concurrency import MaxConcurrency as MaxConcurrency
