+++ src/core/checkers/max_concurrency.py
@@ -5,15 +5,13 @@ import logging
 from collections import deque
 from typing import TYPE_CHECKING, Any, Callable, Deque, Hashable, TypeVar, Union
 
-import discord
-from discord.app_commands import Command, ContextMenu, check as app_check
 from typing_extensions import Self
 
-from .errors import BotMissingPermissions, MaxConcurrencyReached
-from .misc_command import MiscCommandContext, misc_check as misc_check
-from .utils import CommandType
+from ..errors import MaxConcurrencyReached
+from ..misc_command import misc_check as misc_check
 
 T = TypeVar("T")
+
 
 logger = logging.getLogger(__name__)
 
@@ -21,9 +19,7 @@ logger = logging.getLogger(__name__)
 if TYPE_CHECKING:
     from discord import Interaction
 
-    from mybot import MyBot
-
-    from ._types import CoroT
+    from .._types import CoroT
 
     MaxConcurrencyFunction = Union[Callable[[Interaction], CoroT[T]], Callable[[Interaction], T]]
 
@@ -136,92 +132,3 @@ class _Semaphore:
     def release(self) -> None:
         self.value += 1
         self.wake_up()
-
-
-def add_extra(type_: CommandType, func: T, name: str, value: Any) -> T:
-    copy_func = func  # typing behavior
-    if type_ is CommandType.APP:
-        if isinstance(func, (Command, ContextMenu)):
-            func.extras[name] = value
-        else:
-            logger.critical(
-                "Because we need to add extras, this decorator must be above the command decorator. "
-                "(Command should already be defined)"
-            )
-    elif type_ is CommandType.MISC:
-        if hasattr(func, "__listener_as_command__"):
-            command: Command[Any, ..., Any] = getattr(func, "__listener_as_command__")
-            command.extras[name] = value
-        else:
-            if not hasattr(func, "__misc_commands_extras__"):
-                setattr(func, "__misc_commands_extras__", {})
-            getattr(func, "__misc_commands_extras__")[name] = value
-    return copy_func
-
-
-def bot_required_permissions_predicate(perms: dict[str, bool]) -> Callable[..., bool]:
-    def predicate(ctx: Interaction | MiscCommandContext[MyBot]):
-        match ctx:
-            case discord.Interaction():
-                permissions = ctx.app_permissions
-            case MiscCommandContext():
-                permissions = ctx.bot_permissions
-
-        missing = [perm for perm, value in perms.items() if getattr(permissions, perm) != value]
-
-        if not missing:
-            return True
-
-        raise BotMissingPermissions(missing)
-
-    return predicate
-
-
-def _bot_required_permissions(type_: CommandType, **perms: bool) -> Callable[[T], T]:
-    invalid = set(perms) - set(discord.Permissions.VALID_FLAGS)
-    if invalid:
-        raise TypeError(f"Invalid permission(s): {', '.join(invalid)}")
-
-    def decorator(func: T) -> T:
-        match type_:
-            case CommandType.APP:
-                add_extra(
-                    type_, func, "bot_required_permissions", [perm for perm, value in perms.items() if value is True]
-                )
-                return app_check(bot_required_permissions_predicate(perms))(func)
-            case CommandType.MISC:
-                add_extra(
-                    CommandType.MISC,
-                    func,
-                    "bot_required_permissions",
-                    [perm for perm, value in perms.items() if value is True],
-                )
-                return misc_check(bot_required_permissions_predicate(perms))(func)
-
-    return decorator
-
-
-def misc_cmd_bot_required_permissions(**perms: bool) -> Callable[[T], T]:
-    invalid = set(perms) - set(discord.Permissions.VALID_FLAGS)
-    if invalid:
-        raise TypeError(f"Invalid permission(s): {', '.join(invalid)}")
-    return _bot_required_permissions(CommandType.MISC, **perms)
-
-
-def app_command_bot_required_permissions(**perms: bool) -> Callable[[T], T]:
-    invalid = set(perms) - set(discord.Permissions.VALID_FLAGS)
-    if invalid:
-        raise TypeError(f"Invalid permission(s): {', '.join(invalid)}")
-    return _bot_required_permissions(CommandType.APP, **perms)
-
-
-async def is_user_authorized(context: MiscCommandContext[MyBot]) -> bool:
-    del context  # unused
-    # TODO(airo.pi_): check using the database if the user is authorized
-    return True
-
-
-async def is_activated(context: MiscCommandContext[MyBot]) -> bool:
-    del context  # unused
-    # TODO(airo.pi_): check using the database if the misc command is activated
-    return True
